feat: Refactor order actions to separate submit and cancel functionalities into distinct modules
